rename customTypes.ts -> interfaces.ts
